requirements.txt
streamlit==1.31.1
pandas==2.2.0
pytest==8.0.0
PyYAML==6.0.1